post

+++ app/Http/Controllers/TasksController.php
@@ -15,9 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::user()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('tasks.index',['tasks' => $tasks,]);
+        return view('welcome',$data);
         
     }
 
@@ -43,14 +52,13 @@ class TasksController extends Controller
     {
         $this->validate($request, [
             'content'=> 'required|max:191',
-            'status' => 'required|max:10'
         ]);
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
         
-        return redirect('/');
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+        ]);
+        
+        return back();
     }
 
     /**
@@ -108,9 +116,14 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
         
-        return redirect('/');
+        print $task->user_id;
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
+        return back();
     }
 }
